Add @routes custom Blade directive
The @routes is a custom Blade directive, provided by the Ziggy package,
rather than a native Blade directive.

When we include @routes in the Blade template, typically in the <head>
section, Ziggy uses it to inject a JSON representation of the Laravel
application's named routes into the frontend as a JavaScript object.

This allows us to use Laravel-style named routes in the JavaScript or
TypeScript code, especially useful in SPAs built with Inertia.js, Vue,
React, etc.