Install ziggy-js package
Ziggy registers a Blade directive named routes.  When Laravel renders
the Blade view, @routes is compiled to something like this:

<script>
    window.Ziggy = {"url":"http://localhost","port":null,"defaults":{},"routes":{...}};
</script>
<script src="/vendor/ziggy/js/route.js"></script>

The route() helper function becomes available globally in JavaScript or
TypeScript.  We must also install the Ziggy frontend package.  Since we
use Vite, we need to import it explicitly.

$ sail npm install ziggy-js

added 2 packages, and audited 304 packages in 2s